refactoring: splitting up core package into multiple packages
